+++ src/semantic/src/scopes/index.ts
@@ -1,6 +1,7 @@
 // Public API: expose a single public coordinator type for external consumers.
 export * from "./identifier-scope.js";
 export * from "./occurrence.js";
+export * from "./path-sorting.js";
 export * from "./scope-override-keywords.js";
 export * from "./types.js";
